Rename viewModel
